backend/requirements/base.txt
# Django及相关框架
Django==4.2.7
djangorestframework==3.14.0

# 数据库驱动
mysqlclient==2.2.0

# 认证相关
djangorestframework-simplejwt==5.3.0

# 跨域支持
django-cors-headers==4.3.1

# 环境变量管理
python-decouple==3.8

# 时区处理
pytz==2023.3

# 数据验证
pydantic==2.5.0